requirements.txt
pyTelegramBotAPI==4.21.0
openai==1.51.0
python-dotenv==1.0.1
Flask==3.0.3
